requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
Pillow==2.9.0
PyJWT==1.4.0
Pygments==2.0.2
South==1.0.2
Unidecode==0.04.18
django-cache-utils==0.7.2
django-filter==0.10.0
django-jsonfield==0.9.13
django-nvd3==0.8.2
django-qsstats-magic==0.7.2
django-suit==0.3a1
django-suit-redactor==0.0.2
djangorestframework==3.1.3
djangorestframework-jwt==1.6.0
httpie==0.9.2
python-dateutil==2.4.2
python-memcached==1.54
python-nvd3==0.13.7
python-slugify==0.0.9
requests==2.7.0
six==1.9.0
cloudinary
django-rest-swagger
django-crontab
sorl-thumbnail
django-push-notifications
django-markupfield
